requirements.txt
absl-py==2.1.0
altgraph==0.17.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
bleak==0.22.2
bleak-winrt==1.2.0
bottle==0.13.4
branca==0.7.2
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
evillimiter==1.5.0
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
folium==0.17.0
fonttools==4.56.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.7
instaloader==4.14.1
ipython==8.12.3
iso8601==2.1.0
ixml==0.1.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
netaddr==1.3.0
netifaces==0.11.0
networking==0.0.1
numpy==2.0.1
opt_einsum==3.4.0
optree==0.14.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib2==2.3.7.post1
pefile==2023.2.7
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
proxy-tools==0.1.0
pure_eval==0.2.3
py_positron==0.0.2.3
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.21.0
pyglet==2.1.6
Pygments==2.19.1
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.2.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.5
pytz==2024.1
pywebview==5.4
pywin32==308
pywin32-ctypes==0.2.2
PyYAML==6.0.2
pyzmq==26.2.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scapy==2.6.1
scikit-learn==1.5.1
scipy==1.14.0
serial==0.0.97
simplepyble==0.7.3
six==1.16.0
soupsieve==2.5
speedtest-cli==2.1.3
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tinycss2==1.4.0
tk==0.1.0
tkinterweb==3.25.17
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xyzservices==2024.6.0
yarg==0.1.9
zope.interface==6.4.post2
